Match calendar event cards to events page design (volunteer + admin)
