+++ src/components/AdminEventsPage.tsx
@@ -1,72 +1,8 @@
 "use client";
 import { useState } from "react";
-import { useRouter } from "next/navigation";
 import styles from "@/styles/AdminEventsPage.module.css";
-import { AppEvent, MOCK_EVENTS } from "@/data/events";
-
-function EventCard({ event }: { event: AppEvent }) {
-  const monthLabel = event.date.toLocaleString("en-US", { month: "long" });
-  const dayNumber = event.date.getDate();
-  const router = useRouter();
-
-  return (
-    <div className={styles.card}>
-      <div className={styles.cardBg} style={{ backgroundImage: `url(${event.imageUrl})` }} />
-      <div className={styles.cardOverlay} />
-
-      <div className={styles.admin}>
-        <div className={styles.registered}>Registered: {event.registeredCount}</div>
-        {event.section === "past" && (
-          <div className={styles.attendance}>
-            Attended: {event.attendanceCount ?? "—"}/{event.registeredCount}
-          </div>
-        )}
-      </div>
-
-      <div className={styles.cardContent}>
-        <div className={styles.cardDate}>
-          <div className={styles.cardMonth}>{monthLabel}</div>
-          <div className={styles.cardDay}>{dayNumber}</div>
-        </div>
-
-        <div className={styles.cardLocation}>
-          <span className={styles.pin}>📍</span>
-          <span className={styles.school}>{event.school}</span>
-        </div>
-      </div>
-
-      <div className={styles.hoverPanel}>
-        <div className={styles.hoverPanelInner}>
-          <div className={styles.hoverTitle}>
-            {event.date.toLocaleDateString("en-US", {
-              weekday: "long",
-              month: "long",
-              day: "numeric",
-            })}
-          </div>
-          <div className={styles.hoverText}>
-            {event.startTime} - {event.endTime}
-          </div>
-          <div className={styles.hoverText}>{event.school}</div>
-
-          <div className={styles.hoverButtons}>
-            <button type="button" className={styles.hoverBtnLight}>
-              Edit Event
-            </button>
-
-            <button
-              type="button"
-              className={styles.hoverBtnDark}
-              onClick={() => router.push(`/admin-events/${event.id}`)}
-            >
-              {event.section === "past" ? "Event Report" : "View Event Info"}
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
+import AdminEventCard from "@/components/AdminEventCard";
+import { MOCK_EVENTS } from "@/data/events";
 
 function DateRangeControl({
   start,
@@ -84,6 +20,7 @@ function DateRangeControl({
     const [y, m, d] = value.split("-").map(Number);
     return new Date(y, m - 1, d);
   };
+
   return (
     <div className={styles.dateRow}>
       <input
@@ -135,7 +72,7 @@ export default function AdminEventsPage() {
 
         <div className={styles.row}>
           {upcomingSorted.map((event) => (
-            <EventCard key={event.id} event={event} />
+            <AdminEventCard key={event.id} event={event} />
           ))}
         </div>
       </section>
@@ -147,7 +84,7 @@ export default function AdminEventsPage() {
 
         <div className={styles.row}>
           {pastSorted.map((event) => (
-            <EventCard key={event.id} event={event} />
+            <AdminEventCard key={event.id} event={event} />
           ))}
         </div>
       </section>
